refactor: split piAgentManager into session, model, lifecycle, and turn modules
Keep piAgentManager as the public API tests and main.ts import, and move
session persistence, model registration, session reuse, and turn execution
into focused sibling files.
